Handle more edge cases for Delete()
We should never merge into the root node since it's assumed to be the
empty string.

We need to check for deleting this node from the parent before we have
potentially merged the node and its child.  If we check after we've
merged, it might be the case that the merged node has no children (in
which case we don't want to delete it from the parent, since it's still
a valid entry.)

This patch includes tests to check that we're handling these cases.