Allow calling Delete() inside Walk() and WalkPrefix()
In recursiveWalk called by Walk and WalkPrefix, we were iterating over
the edges of a node. The edges as well as the current node object itself
(when mergeChild is called) can be modified by calls to Delete.

This commit modifies the loop to handle modifications correctly and
prevent a nil pointer dereference panic in the case detailed in TestWalkDelete.

Signed-off-by: Tibor Vass <teabee89@gmail.com>